Modify HTML and JS
Modified index.html to add an h1 element to the page as a placeholder.
Modified script.js to add an array to store books, an object constructor to make book objects with title, author, page count, and if the book has been read, added a function to add
books to the myLibrary array, and added Crime and Punishment as a manual entry for testing purposes.