Modify HTML, CSS, and JS
Modified index.html to add the form so that users can add books.
Modified script.js to make the 'add book' button show a dialog with an HTML form in it so users can add books.
Modified style.css to add styling to the buttons of the page and the HTML form so it looks cleaner. Still needs some work.