Modify HTML, CSS, and JS
Modified index.html to add placeholder text for when the user has not added a book yet.
Modified script.js to check if a new book already exists in the array using the index.some() function and altered the getBooks() function to clear the list each time it is called
so that there were no duplicate items when someone submitted the add book form.