Modify JS and CSS
Modified script.js to add a function that marks a book as read and then updates the book cards. There may be some redundancy in the markAsRead and getBooks functions but oh well.
Modified style.css to add a class for the mark as read button to become 'disabled' with the has-been-read-btn class.